-Clean and Categorize code
