+++ src/Covid19/UI.java
@@ -32,7 +32,7 @@ public class UI {
 
             switch (DataChoice) {
                 case 1 -> ReadInclusiveDateRange.main();
-                case 2 -> DaysFromAParticularDate.main();
+                case 2 -> DaysFromAParticularDate.Process();
                 case 3 -> DaysToAParticularDate.main();
             }
         }
